refactor(app): move runtime code to src
