+++ app/__tests__/App.bootstrap.test.tsx
@@ -1,4 +1,4 @@
-/* eslint-disable import/order, @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import fs from 'fs';
 import path from 'path';
 
@@ -7,7 +7,7 @@ import React from 'react';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import renderer from 'react-test-renderer';
 
-import App from '../App';
+import App from '../src/App';
 
 jest.mock('expo-notifications', () => ({
   getPermissionsAsync: (jest.fn() as any).mockResolvedValue({ status: 'granted' }),
@@ -17,12 +17,12 @@ jest.mock('expo-notifications', () => ({
   cancelScheduledNotificationAsync: jest.fn() as any,
 }));
 
-jest.mock('../features/Habits/components/GoalModal', () => () => null);
-jest.mock('../features/Habits/components/HabitSettingsModal', () => () => null);
-jest.mock('../features/Habits/components/MissedDaysModal', () => () => null);
-jest.mock('../features/Habits/components/OnboardingModal', () => () => null);
-jest.mock('../features/Habits/components/ReorderHabitsModal', () => () => null);
-jest.mock('../features/Habits/components/StatsModal', () => () => null);
+jest.mock('../src/features/Habits/components/GoalModal', () => () => null);
+jest.mock('../src/features/Habits/components/HabitSettingsModal', () => () => null);
+jest.mock('../src/features/Habits/components/MissedDaysModal', () => () => null);
+jest.mock('../src/features/Habits/components/OnboardingModal', () => () => null);
+jest.mock('../src/features/Habits/components/ReorderHabitsModal', () => () => null);
+jest.mock('../src/features/Habits/components/StatsModal', () => () => null);
 jest.mock('react-native-emoji-selector', () => 'EmojiSelector');
 
 describe('App bootstrap', () => {
@@ -32,7 +32,7 @@ describe('App bootstrap', () => {
   });
 
   it('imports react-native-reanimated before app bootstrap', () => {
-    const indexPath = path.join(__dirname, '..', 'index.ts');
+    const indexPath = path.join(__dirname, '..', 'src', 'index.ts');
     const content = fs.readFileSync(indexPath, 'utf8');
     expect(content.trim().startsWith("import 'react-native-reanimated'")).toBe(true);
   });
