+++ app/jest.config.js
@@ -4,6 +4,11 @@ module.exports = {
   preset: 'react-native',
   testEnvironment: 'node',
   setupFilesAfterEnv: ['@testing-library/jest-native/extend-expect'],
+  roots: ['<rootDir>/src', '<rootDir>/__tests__'],
+  moduleDirectories: ['node_modules', 'src'],
+  moduleNameMapper: {
+    '^@/(.*)$': '<rootDir>/src/$1',
+  },
   transformIgnorePatterns: [
     'node_modules/(?!(' +
       'react-native|' +
