+++ app/src/constants/stageColors.ts
