+++ app/src/features/Habits/HabitsScreen.tsx
@@ -7,8 +7,8 @@ import { Alert, FlatList, Text, TouchableOpacity, View, Modal } from 'react-nati
 import EmojiSelector from 'react-native-emoji-selector';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
-import { spacing } from '../../Sources/design/DesignSystem';
-import useResponsive from '../../Sources/design/useResponsive';
+import { spacing } from '../../../Sources/design/DesignSystem';
+import useResponsive from '../../../Sources/design/useResponsive';
 
 import GoalModal from './components/GoalModal';
 import HabitSettingsModal from './components/HabitSettingsModal';
