+++ app/src/features/Journal/JournalScreen.tsx
