+++ app/src/features/Map/__tests__/stageData.test.ts
